Keep a scroll's map offset signed so a narrow room cannot wrap it (B19)
Walking south from Deepwood Shrine's pot-bridge room into the double-statue
room crashed the game outright. Reported from Android with a --record script
and a logcat, and it reproduced on desktop on the first replay -- SIGSEGV, exit
139, deterministic, in a debugger. Two Android-only reports in this milestone
cost rounds of asking what was different about the device; this one cost one
replay, which is what the diagnostics in the previous commits were for.

**Cause.** sub_0807D280 case 2 -- the southward branch -- computed

    mapspecial + (((unk_18 * 0x10000 >> 0x12) << 8) + ((xdiff >> 4) << 1))

unk_18 is the function's only u32 local, and mixing it into the sum makes the
whole expression unsigned by the usual arithmetic conversions. A negative xdiff
then does not subtract, it wraps. Measured at the fault: unk_18 = 0 and
xdiff = -24 give 0u + (-4) = 0xFFFFFFFC, and the pointer advanced 4,294,967,292
entries -- 8.6 GB -- out of a 32 KB array. The DmaSet on the next line is the
memcpy that died.

xdiff is negative exactly when the room is narrower than the viewport, because
the camera is then pinned left of the room origin (VIEWPORT_CAM_MIN_X) and the
columns either side are backdrop. At GBA-native width that is unreachable: 240
is also the narrowest room in the game, so xdiff >= 0 always and the unsigned
sum was free. At 320 it is the common case -- 443 of 617 rooms are narrower,
and this one is 272. The other three branches build their offsets out of the
s32 tmp locals and were already signed; case 2 is the only one that needed it.

**Fix.** Cast the u32 term to s32 so the sum is signed. At GBA-native size both
spellings agree exactly, because the sum cannot be negative there.

**Evidence.** The reporter's recording segfaults at frame ~8161 before and runs
to completion after -- exit 0 over 8400 frames, verified through the installed
play binary as well as build/pc. Regression gate at 240x160: canonical route
11/11 with 0 differences, map-source audit 0 mismatched in 265,497,600 fetches.

Same shape as the carried-forward 8-bit window masks: "at 240 a screen x could
not exceed 255 and the mask was free" is the same sentence as "at 240 a camera
offset could not be negative and the unsigned type was free". Recorded as
lesson 17, with the sibling case the fix deliberately does not touch -- case 1
and the default branch feed a negative ydiff to (ydiff >> 4) * 0x100, which is
signed and so reads *before* the map rather than past it. It does not crash and
nothing has reproduced it, so it is logged in milestone2-status.md rather than
changed here.

+++ src/screenTileMap.c
@@ -54,7 +54,27 @@ void sub_0807D280(u16* mapspecial, u16* bgbuffer) {
                 DmaSet(3, bgbuffer + 0x40, bgbuffer, 0x800003c0);
             }
 
-            mapspecial = mapspecial + (((unk_18 * 0x10000 >> 0x12) << 8) + ((xdiff >> 4) << 1));
+            /* The cast keeps this sum *signed*. `unk_18` is the only u32 local
+             * in the function, and mixing it into the sum makes the whole
+             * expression unsigned by the usual arithmetic conversions — so a
+             * negative `xdiff` does not subtract, it wraps. Measured at the
+             * crash: unk_18 = 0 and xdiff = -24 gave 0u + (-4) = 0xFFFFFFFC,
+             * and the pointer advanced 4294967292 entries — 8.6 GB — straight
+             * out of gMapDataBottomSpecial.
+             *
+             * `xdiff` is negative whenever the room is narrower than the
+             * viewport, because the camera is then pinned left of the room
+             * origin (VIEWPORT_CAM_MIN_X) and the border either side is
+             * backdrop. At GBA-native width that cannot happen — 240 is also
+             * the narrowest room in the game, so xdiff >= 0 always and the
+             * unsigned sum was free. At 320 it is the common case: 443 of 617
+             * rooms are narrower.
+             *
+             * The other three cases build their offsets out of the s32 tmp
+             * locals and are already signed; this is the only one that needed
+             * it. At GBA-native size both spellings agree exactly, because the
+             * sum cannot be negative there. B19. */
+            mapspecial = mapspecial + ((s32)((unk_18 * 0x10000 >> 0x12) << 8) + ((xdiff >> 4) << 1));
             DmaSet(3, mapspecial, bgbuffer + 0x280, 0x80000020);
             DmaSet(3, mapspecial + 0x80, bgbuffer + 0x2a0, 0x80000020);
             return;
